refactor:TaskMaster remove option int cast

+++ ExtremeRoles/Roles/Solo/Neutral/TaskMaster.cs
@@ -120,19 +120,19 @@ namespace ExtremeRoles.Roles.Solo.Neutral
             CustomOptionBase parentOps)
         {
             CustomOption.Create(
-                this.GetRoleOptionId((int)TaskMasterOption.AddCommonTaskNum),
+                this.GetRoleOptionId(TaskMasterOption.AddCommonTaskNum),
                 string.Concat(
                     this.RoleName,
                     TaskMasterOption.AddCommonTaskNum.ToString()),
                 1, 0, 15, 1, parentOps);
             CustomOption.Create(
-                this.GetRoleOptionId((int)TaskMasterOption.AddLongTaskNum),
+                this.GetRoleOptionId(TaskMasterOption.AddLongTaskNum),
                 string.Concat(
                     this.RoleName,
                     TaskMasterOption.AddLongTaskNum.ToString()),
                 1, 0, 15, 1, parentOps);
             CustomOption.Create(
-                this.GetRoleOptionId((int)TaskMasterOption.AddNormalTaskNum),
+                this.GetRoleOptionId(TaskMasterOption.AddNormalTaskNum),
                 string.Concat(
                     this.RoleName,
                     TaskMasterOption.AddNormalTaskNum.ToString()),
@@ -143,11 +143,11 @@ namespace ExtremeRoles.Roles.Solo.Neutral
         {
             var allOption = OptionHolder.AllOption;
             this.addLongTask = allOption[
-                GetRoleOptionId((int)TaskMasterOption.AddLongTaskNum)].GetValue();
+                GetRoleOptionId(TaskMasterOption.AddLongTaskNum)].GetValue();
             this.addNormalTask = allOption[
-                GetRoleOptionId((int)TaskMasterOption.AddNormalTaskNum)].GetValue();
+                GetRoleOptionId(TaskMasterOption.AddNormalTaskNum)].GetValue();
             this.addCommonTask = allOption[
-                GetRoleOptionId((int)TaskMasterOption.AddCommonTaskNum)].GetValue();
+                GetRoleOptionId(TaskMasterOption.AddCommonTaskNum)].GetValue();
             this.addTask.Clear();
         }
